feat: added a general `run_mcp_and_devices.py` script for starting Tango DB for a specific device and all of its sub devices, deleted `start_mcp_with_twin.py`
